updated the Versionnumber
besides the versionnumber the Welcome text could use an update, too.
I didnt want to change too much but i just loaded the gummi-gtk3 ppa to ubuntu and saw the old Version named in the Welcometext while the programinfo said its 7.1.

The Projectsite is down, too. So there might need some more to be sorted.